rename file and remove mooncake cruft
